requirements.txt
amqp==2.6.0
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-celery-beat==2.0.0
django-cors-headers==3.3.0
django-filter==2.4.0
django-jsonfield==1.4.0
django-picklefield==2.1.1
django-timezone-field==4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==19.9.0
idna==2.9
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.9
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
psycopg2==2.8.3
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
